Версия 08.02.11 Peaлизовано:         - все индексы переведены в unsigned long или unsigned int типы         - почищены  warnings компилятора (36)         - избавиться от глобальных констант - введён соотв. статический singleton Исправлено:         - исправлена ошибка в                 void Task::Control_AltInt(int trNum)                 void Task::Control_Puntryagin(int trNum)                 void Task::Control_R1(int trNum)                 void Task::Control_R2(int trNum)           неправильный расчёт индекса по основному циклу         - исправлена ошибка в расчёте хода времени в методах                 void PR_Task::Control_PR(int trNum)                 void PR_Task::Control_PR_fullSets(int trNum)           некорректное условие расчёта при оценке не прошли ли верхнюю границу множества управляемости         - вывод времени в файл в задачах уклонения

+++ pursuerType.cpp
@@ -8,7 +8,7 @@
 
 
 //----------------------------------------------------------------------------
-pursuerType::pursuerType(int d, short p, long s):
+pursuerType::pursuerType( unsigned long d, short p,unsigned long s):
 	dim(d), perfomance(p), steps(s){
 
 }
@@ -22,10 +22,6 @@ pursuerType::~pursuerType() {
 istream& __fastcall operator >> (istream& in_data, pursuerType& C) {
 	char c;
 	char long_buf[2048];
-	//double tempVal;
-	//long i, j; // ,k=0,l=0;
-   //	string *tmpStr;
-   //	TNetF* tmpPRNet;
 	Vector *xs;
 
 		in_data.get(c);
@@ -58,12 +54,9 @@ istream& __fastcall operator >> (istream& in_data, pursuerType& C) {
 			in_data.get(c);
 			cout << *xs;
 		}while ((c != ';'));
-	  //	in_data.get(c);
- //	}
 	do{
 		in_data.get(c);
 	}while ((c != '\n')&&(!in_data.eof()));
-   //	in_data.get(c);
 
 	return in_data;
 }
